Cetegorize files into directories
