+++ XcodeServerSDK/Server Entities/Bot.swift
