+++ XcodeServerSDK/Server Entities/Device.swift
