+++ XcodeServerSDK/Server Entities/Integration.swift
