+++ XcodeServerSDK/Server Helpers/Errors.swift
