+++ XcodeServerSDK/Utils/Logging.swift
